requirements.txt
anthropic==0.36.0
APScheduler==3.10.4
chromadb==0.4.14
email-reply-parser==0.5.12
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
IMAPClient==3.0.1
InstructorEmbedding==1.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
pyzmail36==1.0.5
# langchain==0.0.133
openai==0.28.1
sentence-transformers==2.2.2
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.140.0
SQLAlchemy==2.0.32
# talon==1.4.4
torch==2.4.0
uuid6==2023.5.2